+++ src/context.jsx
@@ -1,4 +1,3 @@
 import { createContext } from 'react';
 
 export const PostContext = createContext();
-export const FormContext = createContext();
